Refactor options to JS modules
