+++ src/js/_share/utils.js
@@ -87,3 +87,10 @@ export function getPluralForm(pluralCount = 1, translatedString = '') {
       return translatedString;
   }
 }
+
+export function formatByteSize(bytes) {
+	if(bytes < 1024) return bytes + " bytes";
+	else if(bytes < 1048576) return(bytes / 1024).toFixed(3) + " KiB";
+	else if(bytes < 1073741824) return(bytes / 1048576).toFixed(3) + " MiB";
+	else return(bytes / 1073741824).toFixed(3) + " GiB";
+};
