+++ src/js/options/theme.js
@@ -1,0 +1,5 @@
+export async function saveOptionTheme() {
+  await browser.storage.sync.set({
+    theme: document.querySelector('input[name="theme"]:checked').value
+  });
+}
