+++ src/js/options/toolbar.js
@@ -1,0 +1,5 @@
+export async function saveOptionToolbarPosition() {
+  await browser.storage.sync.set({
+    toolbarPosition: document.querySelector('input[name="toolbarPosition"]:checked').value
+  });
+}
