+++ src/options.html
@@ -3,6 +3,7 @@
 	<head>
 		<meta charset="utf-8">
 		<link rel="stylesheet" type="text/css" href="css/options.css">
+		<script type="module" src="js/options/index.js"></script>
 	</head>
 	<body>
 		<section id="optionKeyboardShortcuts">
@@ -125,7 +126,5 @@
 				<strong><span id="thumbnailCacheSize"></span></strong>
 			</div>
 		</section>
-		<script src="js/backup.js"></script>
-		<script src="js/options.js"></script>
 	</body>
 </html>
